backend/requirements.txt
fastapi
uvicorn[standard]
langgraph
langchain-core
langchain-google-genai
google-genai
sqlalchemy
greenlet
aiosqlite
python-dotenv
httpx
pydantic
